Updating onos config loader scripts
When the '>' operand is used for the loader.sh script
it concatenates several lines into the first line of the script.
This prevents the environment variables from being defined
and this script fails.

Changing the CWD to /tmp, since there is no reason to
Write to the /onos/root directory.  This enables the pod
to be run without being a privileged user.